+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `eloquentencryption` will be documented in this file
 
+## 2.0
+- EloquentEncryption now uses `Illuminate\Contracts\Encryption\Encrypter` contract.
+- **BC** If relying on 1.x use `encryptString()` or `decryptString()` functions **if you are using this encryption elsewhere in your application**. As the default encrypt/decrypt function now serialize values automatically, this may cause unexpected errors during decrypting.
+
 ## 1.4
 - Add optional support to define `RsaKeyHandler` to store, retrieved generated RSA keys.